+++ nessus_file_analyzer/dialogs/url_open.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+"""
+nessus file analyzer (NFA) by LimberDuck (pronounced *ˈlɪm.bɚ dʌk*) is a GUI
+tool which enables you to parse multiple nessus files containing the results
+of scans performed by using Nessus by (C) Tenable, Inc. and exports parsed
+data to a Microsoft Excel Workbook for effortless analysis.
+Copyright (C) 2019 Damian Krawczyk
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+"""
+
+from PyQt5.QtCore import QUrl
+from PyQt5.QtGui import QDesktopServices
+
+
+def open_website(url: str):
+    """
+    Open a given URL in the system's default web browser.
+    Returns True if successful, False otherwise.
+    """
+    return QDesktopServices.openUrl(QUrl(url))