+++ CHANGELOG.md
@@ -5,6 +5,24 @@ This document records all notable changes to [nessus file analyzer (NFA) by Limb
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2025-09-01
+
+### Added
+
+- New options:
+  - `Help > Check for Update` - will return confirmation if you are using the latest version of NFA.
+  - `Help > Documentation` - will open NFR documentation at LimberDuck.org.
+  - `Help > GitHub` - will open NFR GitHub page.
+  - `Help > Releases` - will open NFR GitHub Releases page.
+
+- Requirements update
+  - from:
+    - nessus-file-reader>=0.6.0
+  - to:
+    - nessus-file-reader>=0.7.0
+  - new:
+    - packaging>=25.0
+    - requests>=2.32.5
 
 ## [0.8.0] - 2025-06-29
 
@@ -166,6 +184,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
+[0.9.0]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.7.3...v0.8.0
 [0.7.3]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.7.1...v0.7.2